finalizando o cenário de cadastro de food trucks

+++ backend/src/controllers/HelpersController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express'
 
 import User from '../models/User'
+import FoodTruck from '../models/FoodTruck'
 
 export default {
 
@@ -8,6 +9,12 @@ export default {
 
         const { instagram } = req.query
 
+        const user = await User.findOne({instagram: instagram})
+
+        if (!user)
+            return res.status(204).end()
+
+        await FoodTruck.deleteMany({suggested_by: user._id})
         await User.deleteMany({instagram: instagram})
 
         return res.status(204).end()
